flask_backend/requirements.txt
flask
flask-cors
gunicorn
pillow
requests
